refactor(api): split revision router and extract lineage helpers (#232)
