+++ app/api/v1/revision_routes/__init__.py
@@ -1,0 +1,13 @@
+"""Revision child-route modules."""
+
+from app.api.v1.revision_routes.estimates import estimates_router
+from app.api.v1.revision_routes.quantity_takeoffs import (
+    quantity_takeoff_create_router,
+    quantity_takeoffs_router,
+)
+
+__all__ = [
+    "estimates_router",
+    "quantity_takeoff_create_router",
+    "quantity_takeoffs_router",
+]
