+++ app/api/v1/revision_routes/generated_artifacts.py
@@ -1,0 +1,160 @@
+"""Revision generated-artifact read routes."""
+
+from typing import Annotated
+from uuid import UUID
+
+from fastapi import APIRouter, Depends, Query
+from sqlalchemy import select
+from sqlalchemy.ext.asyncio import AsyncSession
+
+from app.api.v1.revision_compat import _get_active_file, _get_active_revision
+from app.api.v1.revision_cursors import (
+    _decode_artifact_cursor,
+    _encode_cursor,
+    _GeneratedArtifactCursor,
+)
+from app.core.exceptions import raise_not_found
+from app.db.session import get_db
+from app.models.file import File
+from app.models.generated_artifact import GeneratedArtifact
+from app.models.project import Project
+from app.schemas.revision import GeneratedArtifactListResponse, GeneratedArtifactRead
+
+generated_artifacts_router = APIRouter()
+
+_DEFAULT_PAGE_SIZE = 50
+_MAX_PAGE_SIZE = 200
+
+
+@generated_artifacts_router.get(
+    "/files/{file_id}/generated-artifacts",
+    response_model=GeneratedArtifactListResponse,
+)
+async def list_file_generated_artifacts(
+    file_id: UUID,
+    db: Annotated[AsyncSession, Depends(get_db)],
+    limit: Annotated[int, Query(ge=1, le=_MAX_PAGE_SIZE)] = _DEFAULT_PAGE_SIZE,
+    cursor: str | None = Query(default=None),
+) -> GeneratedArtifactListResponse:
+    """List active generated artifacts for a file."""
+
+    file = await _get_active_file(file_id, db)
+    if file is None:
+        raise_not_found("File", str(file_id))
+
+    pagination_cursor = _decode_artifact_cursor(cursor) if cursor else None
+
+    query = (
+        select(GeneratedArtifact)
+        .join(
+            File,
+            (File.id == GeneratedArtifact.source_file_id)
+            & (File.project_id == GeneratedArtifact.project_id),
+        )
+        .join(Project, Project.id == GeneratedArtifact.project_id)
+        .where(
+            (GeneratedArtifact.source_file_id == file_id)
+            & (GeneratedArtifact.deleted_at.is_(None))
+            & (File.deleted_at.is_(None))
+            & (Project.deleted_at.is_(None))
+        )
+    )
+
+    if pagination_cursor is not None:
+        query = query.where(
+            (GeneratedArtifact.created_at > pagination_cursor.created_at)
+            | (
+                (GeneratedArtifact.created_at == pagination_cursor.created_at)
+                & (GeneratedArtifact.id > pagination_cursor.id)
+            )
+        )
+
+    result = await db.execute(
+        query.order_by(GeneratedArtifact.created_at.asc(), GeneratedArtifact.id.asc()).limit(
+            limit + 1
+        )
+    )
+    artifacts = result.scalars().all()
+    page = artifacts[:limit]
+    next_cursor = None
+
+    if len(artifacts) > limit and page:
+        last_artifact = page[-1]
+        next_cursor = _encode_cursor(
+            _GeneratedArtifactCursor(
+                created_at=last_artifact.created_at,
+                id=last_artifact.id,
+            )
+        )
+
+    return GeneratedArtifactListResponse(
+        items=[GeneratedArtifactRead.model_validate(artifact) for artifact in page],
+        next_cursor=next_cursor,
+    )
+
+
+@generated_artifacts_router.get(
+    "/revisions/{revision_id}/generated-artifacts",
+    response_model=GeneratedArtifactListResponse,
+)
+async def list_revision_generated_artifacts(
+    revision_id: UUID,
+    db: Annotated[AsyncSession, Depends(get_db)],
+    limit: Annotated[int, Query(ge=1, le=_MAX_PAGE_SIZE)] = _DEFAULT_PAGE_SIZE,
+    cursor: str | None = Query(default=None),
+) -> GeneratedArtifactListResponse:
+    """List active generated artifacts associated with a drawing revision."""
+
+    revision = await _get_active_revision(revision_id, db)
+    if revision is None:
+        raise_not_found("Drawing revision", str(revision_id))
+
+    pagination_cursor = _decode_artifact_cursor(cursor) if cursor else None
+
+    query = (
+        select(GeneratedArtifact)
+        .join(
+            File,
+            (File.id == GeneratedArtifact.source_file_id)
+            & (File.project_id == GeneratedArtifact.project_id),
+        )
+        .join(Project, Project.id == GeneratedArtifact.project_id)
+        .where(
+            (GeneratedArtifact.drawing_revision_id == revision_id)
+            & (GeneratedArtifact.deleted_at.is_(None))
+            & (File.deleted_at.is_(None))
+            & (Project.deleted_at.is_(None))
+        )
+    )
+
+    if pagination_cursor is not None:
+        query = query.where(
+            (GeneratedArtifact.created_at > pagination_cursor.created_at)
+            | (
+                (GeneratedArtifact.created_at == pagination_cursor.created_at)
+                & (GeneratedArtifact.id > pagination_cursor.id)
+            )
+        )
+
+    result = await db.execute(
+        query.order_by(GeneratedArtifact.created_at.asc(), GeneratedArtifact.id.asc()).limit(
+            limit + 1
+        )
+    )
+    artifacts = result.scalars().all()
+    page = artifacts[:limit]
+    next_cursor = None
+
+    if len(artifacts) > limit and page:
+        last_artifact = page[-1]
+        next_cursor = _encode_cursor(
+            _GeneratedArtifactCursor(
+                created_at=last_artifact.created_at,
+                id=last_artifact.id,
+            )
+        )
+
+    return GeneratedArtifactListResponse(
+        items=[GeneratedArtifactRead.model_validate(artifact) for artifact in page],
+        next_cursor=next_cursor,
+    )
